refactor to ts imports and move attestation type
