+++ src/index.ts
@@ -1,0 +1,4 @@
+// run with node --experimental-strip-types
+import './program-config.ts';
+import './program.ts';
+import './types.ts';
